+++ hammerblade/torch/tests/targets.py
@@ -7,4 +7,7 @@
 
 pytest_targets = ["test_hammerblade_device",
                   "test_hammerblade_add_sub",
-                  "test_hammerblade_mul"]
+                  "test_hammerblade_mul",
+                  "test_hammerblade_div",
+                  "test_hammerblade_tensor_manip",
+                  ]
